Add files via upload
Quick update for emulator. This is the version that has the shutdown hook to preserve data that is lost via a crash. It also fixes the issue where the GUI displays the same message twice. The emulator logic (data generation) remains the same so there's no need to reincorporate it. This is more QoL if anything